server5K-fork/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via server5k (pyproject.toml)
aiosignal==1.4.0
    # via aiohttp
asgiref==3.11.0
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-cors-headers
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==25.11.1
    # via daphne
automat==25.4.16
    # via twisted
cbor2==5.7.1
    # via autobahn
cffi==2.0.0
    # via cryptography
channels==4.3.2
    # via
    #   server5k (pyproject.toml)
    #   channels-redis
channels-redis==4.3.0
    # via server5k (pyproject.toml)
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via server5k (pyproject.toml)
django==6.0
    # via
    #   server5k (pyproject.toml)
    #   channels
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-cors-headers==4.9.0
    # via server5k (pyproject.toml)
django-filter==25.2
    # via server5k (pyproject.toml)
djangorestframework==3.16.1
    # via
    #   server5k (pyproject.toml)
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via server5k (pyproject.toml)
drf-spectacular==0.29.0
    # via server5k (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   hyperlink
    #   twisted
    #   yarl
incremental==24.11.0
    # via twisted
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
msgpack==1.1.2
    # via
    #   autobahn
    #   channels-redis
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via incremental
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg==3.3.2
    # via server5k (pyproject.toml)
psycopg-binary==3.3.2
    # via psycopg
py-ubjson==0.16.1
    # via autobahn
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyopenssl==25.3.0
    # via twisted
pyyaml==6.0.3
    # via drf-spectacular
redis==7.1.0
    # via channels-redis
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==14.2.0
    # via server5k (pyproject.toml)
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
service-identity==24.2.0
    # via twisted
sqlparse==0.5.4
    # via django
twisted==25.5.0
    # via daphne
txaio==25.12.2
    # via autobahn
typing-extensions==4.15.0
    # via twisted
tzdata==2025.2
    # via
    #   django
    #   psycopg
ujson==5.11.0
    # via autobahn
uritemplate==4.2.0
    # via drf-spectacular
websockets==15.0.1
    # via server5k (pyproject.toml)
whitenoise==6.11.0
    # via server5k (pyproject.toml)
yarl==1.22.0
    # via aiohttp
zope-interface==8.1.1
    # via twisted
